chatserver/requirements.txt
fastapi
uvicorn
gunicorn
requests
mysql-connector-python
websockets
wsproto

